SAWA - structure/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
